index.py Eugene's part ok. Looks roughly ok imo.
Implemented filter_punctuation for search.py: Guys please check if this affects your parts. IMO it shouldn't, nor will it cause too big of a change in correctness.

Refactored cosine_scores, and therefore also refactored the Rocchio algo which is built-in with cosine_scores, and automatically detects if Rocchio even needs to be executed.

Updated in-code documentation.